Renamed oswel-mark-one to oswel to match proper naming conventions with the project. Fix issue with springboot not starting with running the JAR file. Added server endpoint and jar file location under settings for the python startup process.
Setting this property in the pom.xml <appendAssemblyId>false</appendAssemblyId> to false resulted in the issue with springboot application not starting. However setting it to true resolves the issue, but now creates two jar files where one works, the other (executable) doesn't. Added the welcome page which then transitions to the index page showing the avatar once loaded.